+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Audio.hs
@@ -9,8 +9,8 @@ import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)
 
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.URI (URI)
-import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Duration (Duration)
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI (URI)
+import Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.Duration (Duration)
 import Network.API.Telegram.Bot.Property (Identifiable (Identificator, ident))
 
 data Audio = Audio URI Duration (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Int) deriving Show