+++ Network/API/Telegram/Bot/Object/Update/Message/Content/File/Special/URI.hs
@@ -1,4 +1,4 @@
-module Network.API.Telegram.Bot.Object.Update.Message.Content.File.URI (URI (..)) where
+module Network.API.Telegram.Bot.Object.Update.Message.Content.File.Special.URI (URI (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), ToJSON (toJSON), Value (String), withText)
 import "base" Control.Applicative (pure)